Modification de la popup pour demander à l'utilisateur de rentrer un nom a la fin de la partie | Ajout de commentaires sur certaines lignes

+++ SimonSays/app/src/main/java/com/example/simonsays/ScoreBoard.kt
@@ -9,9 +9,11 @@ import android.os.Bundle
 import android.util.Log
 import android.view.View
 import android.widget.*
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.setPadding
 import kotlinx.android.synthetic.main.activity_score_board.*
+import kotlinx.android.synthetic.main.data_party.*
 import kotlinx.android.synthetic.main.popup_name.view.*
 
 
@@ -23,33 +25,31 @@ class ScoreBoard : AppCompatActivity() {
         setContentView(R.layout.activity_score_board)
         displayPlayerHighscore()
 
-        //btn_remove_all = findViewById(R.id.btn_remove_all)
         btn_remove_all.setOnClickListener {
             removeAll()
         }
 
     }
 
-    private fun removeAll(){
+    private fun removeAll(){    //remove tout le contenue de la bdd
         AppDatabase.get(application).playerDao().deleteAll()
         finish()
         startActivity(intent)
     }
 
-    private fun displayPlayerHighscore(){
+    private fun displayPlayerHighscore(){   //affiche les 10 derniers meilleurs score (ajout dynamiquement)
         var rang = 1
-        //val parent_linear = findViewById<LinearLayout>(R.id.parent_linear)
         AppDatabase.get(application).playerDao().getTenLast().forEach { it ->
 
+            //creation d'un linear layout qui contiendra les infos
             val parent = LinearLayout(this)
-
             parent.layoutParams = LinearLayout.LayoutParams(
                 LinearLayout.LayoutParams.MATCH_PARENT,
                 LinearLayout.LayoutParams.WRAP_CONTENT
             )
             parent.orientation = LinearLayout.HORIZONTAL
 
-
+            //textview qui contient le rang
             val tv1 = TextView(this)
             tv1.text = rang.toString()
             tv1.layoutParams =
@@ -59,6 +59,7 @@ class ScoreBoard : AppCompatActivity() {
             tv1.setPadding(15)
             parent.addView(tv1)
 
+            //textview qui contient le nom
             val tv2 = TextView(this)
             tv2.text = it.name
             tv2.layoutParams =
@@ -68,6 +69,7 @@ class ScoreBoard : AppCompatActivity() {
             tv2.setPadding(15)
             parent.addView(tv2)
 
+            //textview qui contient le score
             val tv3 = TextView(this)
             tv3.text = it.score.toString()
             tv3.layoutParams =
@@ -77,6 +79,7 @@ class ScoreBoard : AppCompatActivity() {
             tv3.setPadding(15)
             parent.addView(tv3)
 
+            //textview qui permet de supprimer la ligne de la bdd
             val tv4 = TextView(this)
             tv4.text = "X"
             tv4.layoutParams =
@@ -87,10 +90,20 @@ class ScoreBoard : AppCompatActivity() {
 
             val playerId = it.id
             tv4.setOnClickListener {
-                AppDatabase.get(application).playerDao().deletePlayerFromId(playerId)
-                Toast.makeText(this, "Le score selectionné a bien été supprimé", Toast.LENGTH_SHORT).show()
-                finish()
-                startActivity(intent)
+                val builder = AlertDialog.Builder(this)
+
+                builder.setTitle("Supprimer?")
+                builder.setMessage("Rang: " + tv1.text + "\nNom: " + tv2.text + "\nScore: " + tv3.text)
+                builder.setCancelable(false)
+
+                builder.setPositiveButton("Oui") { _, _ ->
+                    AppDatabase.get(application).playerDao().deletePlayerFromId(playerId)
+                    Toast.makeText(this, "Le score selectionné a bien été supprimé", Toast.LENGTH_SHORT).show()
+                    finish()
+                    startActivity(intent)
+                }
+                builder.setNegativeButton("Non") { _, _ -> }
+                builder.show()
             }
 
             parent.addView(tv4)
@@ -100,9 +113,5 @@ class ScoreBoard : AppCompatActivity() {
             rang += 1
         }
     }
-
-}
-
-private operator fun Drawable.invoke(s: String) {
 
 }
